Refactor script organization and enhance functionality: update index.html to include module type for dynamic gradient and reorganize script references for clarity. Remove control-panel.js to streamline codebase, and implement color management for lyrics in music-player.js, ensuring real-time updates for user experience. Adjust dynamic-gradient.js for module export compatibility.
