+++ index.html
@@ -245,12 +245,16 @@
     <!-- Toast Container -->
     <div id="toast-container" class="toast-container"></div>
 
-    <script src="js/dynamic-gradient.js"></script>
+    <!-- Dynamic Gradient System -->
+    <script type="module" src="js/dynamic-gradient.js"></script>
     <script src="js/message-bridge.js"></script>
-    <script src="js/control-panel.js"></script>
+    <script src="js/toast-system.js"></script>
+    
+    <!-- Control Panel Modules -->
+    <script type="module" src="js/control-panel/index.js"></script>
+    
     <script src="js/music-player.js"></script>
     <script src="js/video-exporter.js"></script>
-    <script src="js/toast-system.js"></script>
 </body>
 
 </html>
